remove welcome view model
